fix: add Main-empty-look, add favoourites-empty-look
