+++ src/components/header/header.tsx
@@ -35,4 +35,3 @@ function Header(): JSX.Element {
 }
 
 export default Header;
-
