+++ src/pages/main-page/main-page.tsx
@@ -1,15 +1,13 @@
 import { OfferPreview } from '../../mocks/offers';
-import OfferList from '../../components/offer-list/offer-list';
-import { AppRoute } from '../../const';
-import { Link } from 'react-router-dom';
 import Header from '../../components/header/header';
+import CitiesPlaces from '../../components/cities-places/cities-places';
 
 type MainPageProps = {
   offers: OfferPreview[];
 }
 
-function MainPage({offers}: MainPageProps): JSX.Element {
-  const offersCount = offers.length;
+function MainPage({ offers }: MainPageProps): JSX.Element {
+
 
   return (
     <div className='page page--gray page--main'>
@@ -54,33 +52,7 @@ function MainPage({offers}: MainPageProps): JSX.Element {
           </section>
         </div>
         <div className='cities'>
-          <div className='cities__places-container container'>
-            <section className='cities__places places'>
-              <h2 className='visually-hidden'>Places</h2>
-              <b className='places__found'>{offersCount} places to stay in Amsterdam</b>
-              <form className='places__sorting' action='#' method='get'>
-                <span className='places__sorting-caption'>Sort by</span>
-                <span className='places__sorting-type' tabIndex={0}>
-                  Popular
-                  <svg className='places__sorting-arrow' width='7' height='4'>
-                    <use xlinkHref='#icon-arrow-select'></use>
-                  </svg>
-                </span>
-                <ul className='places__options places__options--custom places__options--opened'>
-                  <li className='places__option places__option--active' tabIndex={0}>Popular</li>
-                  <li className='places__option' tabIndex={0}>Price: low to high</li>
-                  <li className='places__option' tabIndex={0}>Price: high to low</li>
-                  <li className='places__option' tabIndex={0}>Top rated first</li>
-                </ul>
-              </form>
-              <div className='cities__places-list places__list tabs__content'>
-                <OfferList offers={offers} />
-              </div>
-            </section>
-            <div className='cities__right-section'>
-              <section className='cities__map map'></section>
-            </div>
-          </div>
+          <CitiesPlaces offers={offers} />
         </div>
       </main>
     </div>
